refactor(render): extract pipeline stages into per-stage modules
Extract the 7 render pipeline stages from the monolithic
RenderPipeline module (1,090 lines) into focused modules:

- RenderPipeline.Scroll (343 lines) — viewport adjustment + buffer fetch
- RenderPipeline.Content (271 lines) — display list draws for windows
- RenderPipeline.Chrome (175 lines) — modeline, minibuffer, overlays
- RenderPipeline.Compose (99 lines) — merge content + chrome into Frame
- RenderPipeline.Emit (68 lines) — protocol encoding + port dispatch

The parent module is now a 196-line orchestrator that calls each
stage module directly.

WindowScroll struct moved into Scroll (its producer). Chrome struct
moved into Chrome (its producer). Co-locating result structs with
their producing modules follows Elixir convention.

Each stage has typed inputs/outputs and its own test file. Shared test
helpers extracted to test/support/render_pipeline_test_helpers.ex.
Integration tests (full pipeline run, dirty-line tracking) stay in
the parent test file since they exercise cross-stage behavior.

Closes #536
