+++ test/minga/editor/render_pipeline/compose_test.exs
@@ -1,0 +1,61 @@
+defmodule Minga.Editor.RenderPipeline.ComposeTest do
+  @moduledoc """
+  Tests for the Compose stage of the render pipeline.
+  """
+
+  use ExUnit.Case, async: true
+
+  alias Minga.Editor.DisplayList.{Cursor, Frame}
+  alias Minga.Editor.Layout
+  alias Minga.Editor.RenderPipeline
+  alias Minga.Editor.RenderPipeline.Chrome
+  alias Minga.Editor.RenderPipeline.Compose
+  alias Minga.Editor.RenderPipeline.Content
+  alias Minga.Editor.RenderPipeline.Scroll
+  alias Minga.Editor.State, as: EditorState
+
+  import Minga.Editor.RenderPipeline.TestHelpers
+
+  # Helper to run through scroll, content, and chrome
+  defp run_through_chrome(state) do
+    state = EditorState.sync_active_window_cursor(state)
+    state = RenderPipeline.compute_layout(state)
+    layout = Layout.get(state)
+    {scrolls, state} = Scroll.scroll_windows(state, layout)
+    {frames, cursor_info, state} = Content.build_content(state, scrolls)
+    chrome = Chrome.build_chrome(state, layout, scrolls, cursor_info)
+    {frames, chrome, cursor_info, state}
+  end
+
+  describe "compose_windows/4" do
+    test "returns a Frame struct" do
+      state = base_state()
+      {frames, chrome, cursor_info, state} = run_through_chrome(state)
+
+      frame = Compose.compose_windows(frames, chrome, cursor_info, state)
+
+      assert %Frame{cursor: %Cursor{}} = frame
+      assert frame.cursor.shape in [:block, :beam, :underline]
+    end
+
+    test "frame windows have modeline injected" do
+      state = base_state()
+      {frames, chrome, cursor_info, state} = run_through_chrome(state)
+
+      frame = Compose.compose_windows(frames, chrome, cursor_info, state)
+
+      [wf | _] = frame.windows
+      assert map_size(wf.modeline) >= 1
+    end
+
+    test "frame includes chrome elements" do
+      state = base_state()
+      {frames, chrome, cursor_info, state} = run_through_chrome(state)
+
+      frame = Compose.compose_windows(frames, chrome, cursor_info, state)
+
+      assert frame.minibuffer != []
+      assert is_list(frame.regions)
+    end
+  end
+end
