+++ test/minga/editor/render_pipeline/scroll_test.exs
@@ -1,0 +1,83 @@
+defmodule Minga.Editor.RenderPipeline.ScrollTest do
+  @moduledoc """
+  Tests for the Scroll stage of the render pipeline.
+  """
+
+  use ExUnit.Case, async: true
+
+  alias Minga.Editor.Layout
+  alias Minga.Editor.RenderPipeline
+  alias Minga.Editor.RenderPipeline.Scroll
+  alias Minga.Editor.RenderPipeline.Scroll.WindowScroll
+  alias Minga.Editor.State, as: EditorState
+
+  import Minga.Editor.RenderPipeline.TestHelpers
+
+  # Helper to run through layout and scroll
+  defp run_through_scroll(state) do
+    state = EditorState.sync_active_window_cursor(state)
+    state = RenderPipeline.compute_layout(state)
+    layout = Layout.get(state)
+    {scrolls, state} = Scroll.scroll_windows(state, layout)
+    {scrolls, state, layout}
+  end
+
+  describe "scroll_windows/2" do
+    test "returns {scrolls, state} for each window" do
+      state = base_state()
+      {scrolls, state, _layout} = run_through_scroll(state)
+
+      assert map_size(scrolls) == 1
+      [{_win_id, scroll}] = Map.to_list(scrolls)
+      assert %WindowScroll{} = scroll
+      assert %EditorState{} = state
+    end
+
+    test "scroll result contains buffer lines" do
+      state = base_state(content: "alpha\nbeta\ngamma")
+      {scrolls, _state, _layout} = run_through_scroll(state)
+      [{_win_id, scroll}] = Map.to_list(scrolls)
+
+      assert "alpha" in scroll.lines
+      assert "beta" in scroll.lines
+      assert "gamma" in scroll.lines
+    end
+
+    test "scroll result has correct cursor at line 0" do
+      state = base_state()
+      {scrolls, _state, _layout} = run_through_scroll(state)
+      [{_win_id, scroll}] = Map.to_list(scrolls)
+
+      assert scroll.cursor_line == 0
+      assert scroll.first_line == 0
+      assert scroll.is_active == true
+    end
+
+    test "gutter_w is non-negative" do
+      state = base_state()
+      {scrolls, _state, _layout} = run_through_scroll(state)
+      [{_win_id, scroll}] = Map.to_list(scrolls)
+
+      assert scroll.gutter_w >= 0
+      assert scroll.content_w >= 1
+    end
+
+    test "scroll result includes buf_version" do
+      state = base_state()
+      {scrolls, _state, _layout} = run_through_scroll(state)
+      [{_win_id, scroll}] = Map.to_list(scrolls)
+
+      assert is_integer(scroll.buf_version)
+      assert scroll.buf_version >= 0
+    end
+
+    test "first frame marks all lines dirty on the window" do
+      state = base_state()
+      {_scrolls, state, _layout} = run_through_scroll(state)
+      [{_win_id, window}] = Map.to_list(state.windows.map)
+
+      # First frame: sentinel values trigger full invalidation
+      assert window.dirty_lines == :all
+    end
+  end
+end
